moved agent and merchant info to different pages
