+++ frontend/fos_tracker/lib/globals.dart
@@ -6,6 +6,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 LatLng startPosition;
 Set<Marker> agentMarkers = new Set();
 Set<Marker> merchantMarkers = new Set();
+Set<Marker> verifiedMerchantMarkers = new Set();
 LatLng currentPosition;
 String pincode;
