+++ frontend/fos_tracker/pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   sliding_up_panel: ^1.0.2
   http: ^0.12.1
   geocoder: ^0.2.1
+  flutter_polyline_points: ^0.1.0
 
 
   # The following adds the Cupertino Icons font to your application.
